+++ src/main/java/com/atlauncher/gui/dialogs/EditModsDialog.java
@@ -34,6 +34,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -55,10 +56,13 @@ import com.atlauncher.App;
 import com.atlauncher.builders.HTMLBuilder;
 import com.atlauncher.data.DisableableMod;
 import com.atlauncher.data.Instance;
+import com.atlauncher.data.ModPlatform;
 import com.atlauncher.data.curseforge.CurseForgeFingerprint;
 import com.atlauncher.data.curseforge.CurseForgeProject;
 import com.atlauncher.data.minecraft.FabricMod;
 import com.atlauncher.data.minecraft.MCMod;
+import com.atlauncher.data.modrinth.ModrinthProject;
+import com.atlauncher.data.modrinth.ModrinthVersion;
 import com.atlauncher.gui.components.ModsJCheckBox;
 import com.atlauncher.gui.handlers.ModsJCheckBoxTransferHandler;
 import com.atlauncher.gui.layouts.WrapLayout;
@@ -70,6 +74,7 @@ import com.atlauncher.network.Analytics;
 import com.atlauncher.utils.CurseForgeApi;
 import com.atlauncher.utils.FileUtils;
 import com.atlauncher.utils.Hashing;
+import com.atlauncher.utils.ModrinthApi;
 import com.atlauncher.utils.Utils;
 
 public class EditModsDialog extends JDialog {
@@ -423,13 +428,77 @@ public class EditModsDialog extends JDialog {
 
                                                 if (curseForgeProject != null) {
                                                     dm.curseForgeProject = curseForgeProject;
-                                                    dm.name = curseForgeProject.name;
-                                                    dm.description = curseForgeProject.summary;
+
+                                                    if (!dm.isFromCurseForge()
+                                                            || App.settings.defaultModPlatform == ModPlatform.CURSEFORGE) {
+                                                        dm.name = curseForgeProject.name;
+                                                        dm.description = curseForgeProject.summary;
+                                                    }
                                                 }
 
                                                 LogManager.debug("Found matching mod from CurseForge called "
                                                         + dm.curseForgeFile.displayName);
                                             });
+                                }
+                            }
+                        }
+                    }
+                }
+
+                if (!App.settings.dontCheckModsOnModrinth) {
+                    Map<String, DisableableMod> sha1Hashes = new HashMap<>();
+
+                    mods.stream()
+                            .filter(dm -> dm.modrinthProject == null && dm.modrinthVersion == null)
+                            .filter(dm -> dm.getFile(instance.ROOT, instance.id) != null).forEach(dm -> {
+                                try {
+                                    sha1Hashes.put(Hashing
+                                            .sha1(dm.disabled ? dm.getDisabledFile(instance).toPath()
+                                                    : dm
+                                                            .getFile(instance.ROOT, instance.id).toPath())
+                                            .toString(), dm);
+                                } catch (Throwable t) {
+                                    LogManager.logStackTrace(t);
+                                }
+                            });
+
+                    if (sha1Hashes.size() != 0) {
+                        Set<String> keys = sha1Hashes.keySet();
+                        Map<String, ModrinthVersion> modrinthVersions = ModrinthApi
+                                .getVersionsFromSha1Hashes(keys.toArray(new String[keys.size()]));
+
+                        if (modrinthVersions != null && modrinthVersions.size() != 0) {
+                            String[] projectIdsFound = modrinthVersions.values().stream().map(mv -> mv.projectId)
+                                    .toArray(String[]::new);
+
+                            if (projectIdsFound.length != 0) {
+                                Map<String, ModrinthProject> foundProjects = ModrinthApi
+                                        .getProjectsAsMap(projectIdsFound);
+
+                                if (foundProjects != null) {
+                                    for (Map.Entry<String, ModrinthVersion> entry : modrinthVersions.entrySet()) {
+                                        ModrinthVersion version = entry.getValue();
+                                        ModrinthProject project = foundProjects.get(version.projectId);
+
+                                        if (project != null) {
+                                            DisableableMod dm = sha1Hashes.get(entry.getKey());
+
+                                            // add Modrinth information
+                                            dm.modrinthProject = project;
+                                            dm.modrinthVersion = version;
+
+                                            if (!dm.isFromModrinth()
+                                                    || App.settings.defaultModPlatform == ModPlatform.MODRINTH) {
+                                                dm.name = project.title;
+                                                dm.description = project.description;
+                                            }
+
+                                            LogManager
+                                                    .debug(String.format(
+                                                            "Found matching mod from Modrinth called %s with file %s",
+                                                            project.title, version.name));
+                                        }
+                                    }
                                 }
                             }
                         }